Removed obsolete properties (slug & count) from mongodb xml-config-files.
Fixes Issue #174.

While doing a composer update a RuntimeException will be thrown because the
property $slug could not be found (Thrown ReflectionClassException).
After removing the slug-line in BaseContext.mongodb.xml and doing a composer
update a "Property $count could not be found" message was thrown.
After removing the count-line in BaseCategory.mongodb.xml and doing a
composer update all will be fine.
These lines have to be removed because the accordingly properties didn't
exist anymore in Model\Category.php and Model\Context.php.